5054 - вернула Long в school id / переименовала name на subjectName

+++ src/main/java/com/kataacademy/schoolportal/common/repository/schooattribute/SchoolRepository.java
@@ -3,6 +3,6 @@ package com.kataacademy.schoolportal.common.repository.schooattribute;
 import com.kataacademy.schoolportal.common.models.schoolatribute.School;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SchoolRepository extends JpaRepository<School, Integer> {
+public interface SchoolRepository extends JpaRepository<School, Long> {
 
 }
